# ARROW-10569: [C++] Improve table filtering performance