@wesm 2020-12-02
Looks good. We could add this benchmark to our ASV suite? Or a C++ benchmark?